Refactor class module names and remove unused imports
