+++ configuration/function_metrics.py
@@ -7,9 +7,9 @@ from typing import Union
 import datetime
 
 from flexeval.classes.Turn import Turn
-from flexeval.classes.Message import Message
+from flexeval.classes.message import Message
 from flexeval.classes.Thread import Thread
-from flexeval.classes.ToolCall import ToolCall
+from flexeval.classes.tool_call import ToolCall
 
 
 ## ~.~ function templates starts ~.~
