+++ src/flexeval/classes/Message.py
@@ -1,20 +1,16 @@
 import copy
 import json
-import os
-import sys
-from pathlib import Path
 
 import peewee as pw
-import pydantic
 from playhouse.shortcuts import model_to_dict
 
 # TODO remove this import from outside of the package
 from configuration import completion_functions
-from flexeval.classes.BaseModel import BaseModel
-from flexeval.classes.Dataset import Dataset
-from flexeval.classes.EvalSetRun import EvalSetRun
-from flexeval.classes.Thread import Thread
-from flexeval.classes.Turn import Turn
+from flexeval.classes.base import BaseModel
+from flexeval.classes.dataset import Dataset
+from flexeval.classes.eval_set_run import EvalSetRun
+from flexeval.classes.thread import Thread
+from flexeval.classes.turn import Turn
 
 
 class Message(BaseModel):
