+++ src/flexeval/classes/Metric.py
@@ -1,23 +1,14 @@
-import copy
-import inspect
-import json
-import os
-import string
-import sys
-from pathlib import Path
-
 import peewee as pw
-import pydantic
 from playhouse.shortcuts import model_to_dict
 
-from flexeval.classes.BaseModel import BaseModel
-from flexeval.classes.Dataset import Dataset
-from flexeval.classes.EvalSetRun import EvalSetRun
-from flexeval.classes.Message import Message
-# from flexeval.classes.DatasetRow import DatasetRow
-from flexeval.classes.Thread import Thread
-from flexeval.classes.ToolCall import ToolCall
-from flexeval.classes.Turn import Turn
+from flexeval.classes.base import BaseModel
+from flexeval.classes.dataset import Dataset
+from flexeval.classes.eval_set_run import EvalSetRun
+from flexeval.classes.message import Message
+
+from flexeval.classes.thread import Thread
+from flexeval.classes.tool_call import ToolCall
+from flexeval.classes.turn import Turn
 
 
 class Metric(BaseModel):
